flatten vault snapshot dir
