+++ src/vault_snapshot/context.py
@@ -4,5 +4,5 @@ from pathlib import Path
 
 # add repo root to sys.path
 # mostly for importing modules in 'lib' dir
-path_root = Path(__file__).parents[2]
+path_root = Path(__file__).parents[1]
 sys.path.append(str(path_root))
